+++ spec/conditions/base_spec.rb
@@ -1,11 +1,9 @@
 # frozen_string_literal: true
 
-require 'test_helper'
-
-describe EightBall::Conditions::Base do
+RSpec.describe EightBall::Conditions::Base do
   describe 'parameter' do
     it 'should return nil' do
-      EightBall::Conditions::Base.new.parameter.must_be_nil
+      expect(EightBall::Conditions::Base.new.parameter).to be_nil
     end
   end
 
@@ -19,17 +17,16 @@ describe EightBall::Conditions::Base do
 
       condition = DummyCondition.new
       condition.parameter = 'someCase'
-      condition.parameter.must_equal 'some_case'
+      expect(condition.parameter).to eq 'some_case'
 
       condition.parameter = 'SomeCase'
-      condition.parameter.must_equal 'some_case'
+      expect(condition.parameter).to eq 'some_case'
     end
   end
 
   describe 'satisfied?' do
     it 'should raise' do
-      e = -> { EightBall::Conditions::Base.new.satisfied? }.must_raise
-      e.message.must_equal 'You can never satisfy the Base condition'
+      expect { EightBall::Conditions::Base.new.satisfied? }.to raise_error RuntimeError, 'You can never satisfy the Base condition'
     end
   end
 end